requirements.txt
flask
requests
pillow
numpy
opencv-python
gunicorn
